content/canvas/test/test_fallback.basic.html

<!DOCTYPE HTML>
<title>Canvas test: fallback.basic</title>
<!-- Testing: Fallback content is inserted into the DOM -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.childNodes.length == 1, "canvas.childNodes.length == 1");

SimpleTest.finish();

});
</script>
